Fixing variable types to match properly the ones that they are assined/compared to.  Some are somewhat serious in 64 bit architectures, e.g. comparisons always rendering false, as in 'src/UtilsParsing.cpp:121:25: warning: comparison is always false due to limited range of data type [-Wtype-limits]'.  This avoids compiler warnings, too.

+++ src/FileParser.cpp
@@ -94,7 +94,7 @@ string FileParser::nextValue() {
 		return ""; // not found
 	}
 	string s;
-	int seppos = val.find_first_of(',');
+	size_t seppos = val.find_first_of(',');
 	if (seppos == string::npos) {
 		s = val;
 		val = "";
